requirements.txt
streamlit==1.31.0
tensorflow-cpu==2.13.0
tf-keras
transformers==4.37.2
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
huggingface-hub==0.20.3
protobuf==3.20.0
streamlit==1.31.0
tensorflow-cpu==2.13.0
tf-keras==2.13.0
transformers==4.37.2
pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
huggingface-hub==0.20.3
protobuf==3.20.0